+++ plone/formwidget/datetime/tests/test_DatetimeFieldWidget.py
@@ -5,7 +5,7 @@ import unittest2 as unittest
 class TestDatetimeFieldWidget(unittest.TestCase):
 
     @mock.patch('plone.formwidget.datetime.z3cform.widget.DatetimeWidget')
-    @mock.patch('plone.formwidget.datetime.z3cform.widget.z3c.form.widget.FieldWidget')
+    @mock.patch('plone.formwidget.datetime.z3cform.widget.FieldWidget')
     def test_DatetimeFieldWidget(self, FieldWidget, DatetimeWidget):
         from plone.formwidget.datetime.z3cform.widget import DatetimeFieldWidget
         field = mock.Mock()