+++ plone/formwidget/datetime/tests/test_MonthYearFieldWidget.py
@@ -5,7 +5,7 @@ import unittest2 as unittest
 class TestMonthYearFieldWidget(unittest.TestCase):
 
     @mock.patch('plone.formwidget.datetime.z3cform.widget.MonthYearWidget')
-    @mock.patch('plone.formwidget.datetime.z3cform.widget.z3c.form.widget.FieldWidget')
+    @mock.patch('plone.formwidget.datetime.z3cform.widget.FieldWidget')
     def test_MonthYearFieldWidget(self, FieldWidget, MonthYearWidget):
         from plone.formwidget.datetime.z3cform.widget import MonthYearFieldWidget
         field = mock.Mock()